requirements.txt
groq==0.4.1
pyyaml==6.0
python-dotenv==1.0.0
pandas==2.0.3
matplotlib==3.7.1
numpy==1.24.3
